Remove custom precompile impls (for now); brush up
