+++ crates/vm2/Cargo.toml
@@ -17,10 +17,6 @@ zkevm_opcode_defs.workspace = true
 zk_evm_abstractions.workspace = true
 primitive-types.workspace = true
 enum_dispatch.workspace = true
-k256.workspace = true
-p256.workspace = true
-sha2.workspace = true
-sha3.workspace = true
 
 # Optional dependencies (used for fuzzing)
 arbitrary = { workspace = true, features = ["derive"], optional = true }
@@ -28,6 +24,7 @@ zk_evm = { workspace = true, optional = true }
 anyhow = { workspace = true, optional = true }
 
 [dev-dependencies]
+assert_matches.workspace = true
 divan.workspace = true
 proptest.workspace = true
 rand.workspace = true
